refactor: rename: StudygroupPostCommentManager, move: 각각의 인터페이스(Manager)
rename: StudygroupPostCommentManager의 중복이름 삭제
move: 각각의 인터페이스Manager를
구현체인 Service 패키지로 이동

[skip ci]
