+++ server/edusync/src/main/java/com/codestates/edusync/model/study/plancalendar/service/CalendarStudygroupService.java
@@ -6,7 +6,6 @@ import com.codestates.edusync.model.member.entity.Member;
 import com.codestates.edusync.model.study.studygroup.entity.Studygroup;
 import com.codestates.edusync.model.study.plancalendar.entity.TimeSchedule;
 import com.codestates.edusync.model.study.plancalendar.repository.CalendarStudygroupRepository;
-import com.codestates.edusync.model.study.plancalendar.utils.CalendarStudygroupManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
