+++ server/edusync/src/main/java/com/codestates/edusync/model/study/studygroupjoin/service/StudygroupJoinService.java
@@ -7,7 +7,6 @@ import com.codestates.edusync.model.member.entity.Member;
 import com.codestates.edusync.model.study.studygroup.entity.Studygroup;
 import com.codestates.edusync.model.study.studygroupjoin.entity.StudygroupJoin;
 import com.codestates.edusync.model.study.studygroupjoin.repository.StudygroupJoinRepository;
-import com.codestates.edusync.model.study.studygroupjoin.utils.StudygroupJoinManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
